Add schema adapation reflecting a given log filter.
This expands the schema reporting to take into account the modifications that a
given log filter imposes on the output. By default, this uses the default
filter, reflecting the field include/exclude set as well as any extension
function.

It also refactors the meat of that analysis into a dedicated function, since the
analyze() event handler was getting lengthy.